Move all AuxKernel and AuxVectorKernel stuff to AuxKernel.C/h
