--- framework/src/auxkernels/AuxVectorKernel.C
@@ -1,19 +1,0 @@
-//* This file is part of the MOOSE framework
-//* https://www.mooseframework.org
-//*
-//* All rights reserved, see COPYRIGHT for full restrictions
-//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
-//*
-//* Licensed under LGPL 2.1, please see LICENSE for details
-//* https://www.gnu.org/licenses/lgpl-2.1.html
-
-#include "AuxVectorKernel.h"
-
-template <>
-InputParameters
-validParams<AuxVectorKernel>()
-{
-  InputParameters params = getAuxKernelTemplValidParams();
-  params.registerBase("AuxVectorKernel");
-  return params;
-}
