+++ framework/src/loops/ComputeNodalAuxBcsThread.C
@@ -14,7 +14,6 @@
 #include "AuxiliarySystem.h"
 #include "FEProblem.h"
 #include "AuxKernel.h"
-#include "AuxVectorKernel.h"
 
 template <typename AuxKernelType>
 ComputeNodalAuxBcsThread<AuxKernelType>::ComputeNodalAuxBcsThread(
