+++ framework/src/loops/ComputeNodalAuxVarsThread.C
@@ -12,7 +12,6 @@
 // MOOSE includes
 #include "AuxiliarySystem.h"
 #include "AuxKernel.h"
-#include "AuxVectorKernel.h"
 #include "FEProblem.h"
 #include "MooseMesh.h"
