+++ framework/src/systems/AuxiliarySystem.C
@@ -11,7 +11,6 @@
 #include "FEProblem.h"
 #include "Factory.h"
 #include "AuxKernel.h"
-#include "AuxVectorKernel.h"
 #include "AuxScalarKernel.h"
 #include "MaterialData.h"
 #include "Assembly.h"
